# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Label Dvisha e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: featured
          in: query
          schema:
            type: boolean
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
    post:
      operationId: createProduct
      tags: [products]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product with full details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductDetail"
        "404":
          description: Not found
    patch:
      operationId: updateProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/images:
    post:
      operationId: addProductImage
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductImage"
  /products/{id}/images/{imageId}:
    delete:
      operationId: deleteProductImage
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/colors:
    post:
      operationId: addProductColor
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductColorInput"
      responses:
        "201":
          description: Color added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductColor"
  /products/{id}/colors/{colorId}:
    delete:
      operationId: deleteProductColor
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: colorId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/sizes:
    post:
      operationId: addProductSize
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductSizeInput"
      responses:
        "201":
          description: Size added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSize"
  /products/{id}/sizes/{productSizeId}:
    patch:
      operationId: updateProductSize
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: productSizeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductSizeUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSize"
    delete:
      operationId: deleteProductSize
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: productSizeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /products/{id}/sections:
    post:
      operationId: addProductSection
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductSectionInput"
      responses:
        "201":
          description: Section added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSection"
  /products/{id}/sections/{sectionId}:
    patch:
      operationId: updateProductSection
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductSectionUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductSection"
    delete:
      operationId: deleteProductSection
      tags: [products]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: sectionId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── REVIEWS ───────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Size:
      type: object
      required: [id, label, sortOrder, isActive]
      properties:
        id:
          type: integer
        label:
          type: string
        bustRange:
          type: ["string", "null"]
        waistRange:
          type: ["string", "null"]
        hipRange:
          type: ["string", "null"]
        heightRange:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
    Category:
      type: object
      required: [id, name, slug, sortOrder, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        cloudinaryPublicId:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isActive:
          type: boolean
...
    ProductImage:
      type: object
      required: [id, productId, imageUrl, sortOrder, isPrimary]
      properties:
        id:
          type: integer
        productId:
          type: integer
        colorId:
          type: ["integer", "null"]
        imageUrl:
          type: string
        cloudinaryPublicId:
          type: ["string", "null"]
        altText:
          type: ["string", "null"]
        sortOrder:
          type: integer
        isPrimary:
          type: boolean
    ProductImageInput:
      type: object
      required: [imageUrl]
      properties:
        colorId:
          type: integer
        imageUrl:
          type: string
        cloudinaryPublicId:
          type: string
        altText:
          type: string
        sortOrder:
          type: integer
        isPrimary:
          type: boolean
    ProductColor:
      type: object
      required: [id, productId, name, hexCode, sortOrder, isAvailable]
      properties:
        id:
          type: integer
        productId:
          type: integer
        name:
          type: string
        hexCode:
          type: string
        sortOrder:
          type: integer
        isAvailable:
          type: boolean
    ProductColorInput:
      type: object
      required: [name, hexCode]
      properties:
        name:
          type: string
        hexCode:
          type: string
        sortOrder:
          type: integer
    ProductSize:
      type: object
      required: [id, productId, sizeId, stockQuantity, isAvailable]
      properties:
        id:
          type: integer
        productId:
          type: integer
        sizeId:
          type: integer
        size:
          $ref: "#/components/schemas/Size"
        stockQuantity:
          type: integer
        isAvailable:
          type: boolean
    ProductSizeInput:
      type: object
      required: [sizeId, stockQuantity]
      properties:
        sizeId:
          type: integer
        stockQuantity:
          type: integer
    ProductSizeUpdate:
      type: object
      properties:
        stockQuantity:
          type: integer
        isAvailable:
          type: boolean
    ProductSection:
      type: object
      required: [id, productId, title, content, sortOrder]
      properties:
        id:
          type: integer
        productId:
          type: integer
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
    ProductSectionInput:
      type: object
      required: [title, content]
      properties:
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
    ProductSectionUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        sortOrder:
          type: integer
    Product:
      type: object
      required: [id, name, slug, sku, price, isActive, isFeatured, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        sku:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        material:
          type: ["string", "null"]
        styleTag:
          type: ["string", "null"]
        deliveryDays:
          type: ["string", "null"]
        allowCustomSize:
          type: boolean
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        lowStockThreshold:
          type: integer
        createdAt:
          type: string
        primaryImage:
          type: ["string", "null"]
        category:
          $ref: "#/components/schemas/Category"
    ProductDetail:
      type: object
      required: [id, name, slug, sku, price, isActive, isFeatured, createdAt, images, colors, sizes, sections]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        shortDescription:
          type: ["string", "null"]
        sku:
          type: string
        price:
          type: number
        compareAtPrice:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        material:
          type: ["string", "null"]
        styleTag:
          type: ["string", "null"]
        deliveryDays:
          type: ["string", "null"]
        allowCustomSize:
          type: boolean
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        lowStockThreshold:
          type: integer
        createdAt:
          type: string
        category:
          $ref: "#/components/schemas/Category"
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProductImage"
        colors:
          type: array
          items:
            $ref: "#/components/schemas/ProductColor"
        sizes:
          type: array
          items:
            $ref: "#/components/schemas/ProductSize"
        sections:
          type: array
          items:
            $ref: "#/components/schemas/ProductSection"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    ProductInput:
      type: object
      required: [name, slug, sku, price]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        sku:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        categoryId:
          type: integer
        material:
          type: string
        styleTag:
          type: string
        deliveryDays:
          type: string
        allowCustomSize:
          type: boolean
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        lowStockThreshold:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        sku:
          type: string
        price:
          type: number
        compareAtPrice:
          type: number
        categoryId:
          type: integer
        material:
          type: string
        styleTag:
          type: string
        deliveryDays:
          type: string
        allowCustomSize:
          type: boolean
        isActive:
          type: boolean
        isFeatured:
          type: boolean
        lowStockThreshold:
          type: integer
    ProductListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer

    # Reviews
    Review:
      type: object
      required: [id, productId, userId, rating, isApproved, createdAt]
      properties:
        id:
          type: integer
        productId:
          type: integer
        userId:
          type: integer
        userName:
          type: ["string", "null"]
        rating:
          type: integer
        title:
          type: ["string", "null"]
        body:
          type: ["string", "null"]
        isApproved:
          type: boolean
        createdAt:
          type: string
...