# lib/api-zod/src/generated/types/productColorInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Dvisha e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductColorInput {
  name: string;
  hexCode: string;
  sortOrder?: number;
}