# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Label Dvisha e-commerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  shortDescription?: string;
  sku: string;
  price: number;
  compareAtPrice?: number;
  categoryId?: number;
  material?: string;
  styleTag?: string;
  deliveryDays?: string;
  allowCustomSize?: boolean;
  isActive?: boolean;
  isFeatured?: boolean;
  lowStockThreshold?: number;
}